Restructure engine file layout
