+++ M2/Macaulay2/e/gb-f4/PolynomialList.hpp
@@ -4,7 +4,7 @@
 #include "../BasicPolyList.hpp"
 #include "MonomialHashTable.hpp"
 #include "MonomialTypes.hpp"
-#include "PolynomialStream.hpp"
+#include "../unused/PolynomialStream.hpp"
 
 namespace newf4 {
