+++ M2/Macaulay2/e/unused/PolynomialStream.hpp
