+++ M2/Macaulay2/e/unused/franzi-brp.cpp
