+++ M2/Macaulay2/e/unused/franzi-gb.cpp
