+++ M2/Macaulay2/e/unused/franzi-interface.cpp
